+++ Sources/DeckKit/DeckViewConfiguration.swift
@@ -6,35 +6,30 @@
 //  Copyright © 2022-2024 Daniel Saidi. All rights reserved.
 //
 
-#if os(iOS) || os(macOS) || os(visionOS)
 import SwiftUI
 
-/**
- This config can be used to configure a ``DeckView``.
- */
+/// This config can be used to configure a ``DeckView``.
 public struct DeckViewConfiguration: Codable, Configuration, Equatable {
 
-    /**
-     Create a stacked deck configuration.
-
-     Note that `alwaysShowLastItem` will make sure that your
-     deck always shows the last item, even if the deck has a
-     large number of items. This makes swiping a card to the
-     back look more consistent, since a card would otherwise
-     fade away as it is swiped to the back of the deck.
-
-     - Parameters:
-       - direction: The visual direction of the stack, by default `.down`.
-       - itemDisplayCount: The max number of items to display, by default `10`.
-       - alwaysShowLastItem: Whether or not to show the last item for visual stability, by default `true`.
-       - scaleOffset: The percentual shrink factor to apply to each item in the stack, by default `0.02`.
-       - verticalOffset: The vertical offset to apply to each item in the stack, by default `10`.
-       - dragRotationFactor: The offset factor with which to rotate an item when it's panned, by default `0.05`.
-       - horizontalDragThreshold: The number of points an item must be panned to be moved to the bottom of the deck, by default `100`.
-       - verticalDragThreshold: The number of points an item must be panned to be moved to the bottom of the deck, by default `250`.
-     */
+    /// Create a deck view configuration.
+    ///
+    /// Note that `alwaysShowLastItem` will make sure that a
+    /// deck always shows the last item in the bottom of the
+    /// stack, even if it has more items. This makes swiping
+    /// away a card look more consistent, since a card would
+    /// otherwise fade away as it is swiped to the back.
+    ///
+    /// - Parameters:
+    ///   - direction: The visual direction of the stack, by default `.down`.
+    ///   - itemDisplayCount: The max number of items to display, by default `10`.
+    ///   - alwaysShowLastItem: Whether or not to show the last item for visual stability, by default `true`.
+    ///   - scaleOffset: The percentual shrink factor to apply to each item in the stack, by default `0.02`.
+    ///   - verticalOffset: The vertical offset to apply to each item in the stack, by default `10`.
+    ///   - dragRotationFactor: The offset factor with which to rotate an item when it's panned, by default `0.05`.
+    ///   - horizontalDragThreshold: The number of points an item must be panned to be moved to the bottom of the deck, by default `100`.
+    ///   - verticalDragThreshold: The number of points an item must be panned to be moved to the bottom of the deck, by default `250`.
     public init(
-        direction: Direction = .down,
+        direction: SwipeDirection = .down,
         itemDisplayCount: Int = 10,
         alwaysShowLastItem: Bool = true,
         scaleOffset: Double = 0.02,
@@ -54,9 +49,13 @@ public struct DeckViewConfiguration: Codable, Configuration, Equatable {
         self.horizontalDragThreshold = horizontalDragThreshold
         self.verticalDragThreshold = verticalDragThreshold
     }
+    
+    public enum SwipeDirection: String, Codable, Sendable {
+        case up, down, left, right
+    }
 
     /// The visual direction of the stack.
-    public var direction: Direction
+    public var direction: SwipeDirection
 
     /// The max number of items to display.
     public var itemDisplayCount: Int
@@ -83,9 +82,7 @@ public struct DeckViewConfiguration: Codable, Configuration, Equatable {
 public extension DeckViewConfiguration {
     
     /// This is a standard deck view configuration.
-    ///
-    /// You can set this value to change the global default.
-    static var standard = Self.init()
+    static var standard: Self { .init() }
 }
 
 public extension DeckViewConfiguration {
@@ -100,4 +97,29 @@ public extension DeckViewConfiguration {
         case up, down
     }
 }
-#endif
+
+public extension View {
+
+    /// Apply a ``DeckViewConfiguration``.
+    func deckViewConfiguration(
+        _ style: DeckViewConfiguration
+    ) -> some View {
+        self.environment(\.deckViewConfiguration, style)
+    }
+}
+
+private extension DeckViewConfiguration {
+
+    struct Key: EnvironmentKey {
+
+        static var defaultValue: DeckViewConfiguration { .standard }
+    }
+}
+
+public extension EnvironmentValues {
+
+    var deckViewConfiguration: DeckViewConfiguration {
+        get { self [DeckViewConfiguration.Key.self] }
+        set { self [DeckViewConfiguration.Key.self] = newValue }
+    }
+}