+++ Sources/DeckKit/_Deprecated/DeckViewConfiguration+Deprecated.swift
@@ -4,8 +4,6 @@ import Foundation
 public extension DeckViewConfiguration {
 
     @available(*, deprecated, message: "This value will be removed in the next major update. Just create a custom configuration when you need to.")
-    static var down = DeckViewConfiguration(
-        direction: .down
-    )
+    static var down: Self { .init(direction: .down) }
 }
 #endif